backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
dj-database-url==2.1.0
psycopg2-binary==2.9.9
Pillow==10.2.0
gunicorn==21.2.0
django-simple-captcha==0.5.20
python-decouple==3.8

django-filter==23.5
